fix favorities state management
