+++ src/lib/dom.ts
@@ -23,5 +23,9 @@ export const listItemsContainer = <HTMLDivElement>document.getElementById('playl
 export const listItemsAnchor = <HTMLAnchorElement>document.getElementById('/list');
 
 export const subtitleContainer = <HTMLElement>document.getElementById('subtitleContainer');
+
 export const subtitleSelector = <HTMLSelectElement>document.getElementById('subtitleSelector');
+
 export const subtitleTrack = <HTMLTrackElement>audio.firstElementChild;
+
+export const favButton = <HTMLInputElement>document.getElementById('favButton');
