+++ src/lib/player.ts
@@ -1,5 +1,5 @@
-import { audio, bitrateSelector, pipedInstances, playButton, subtitleContainer, subtitleSelector, subtitleTrack } from "./dom";
-import { convertSStoHHMMSS, getSaved, params, parseTTML, setMetaData } from "./utils";
+import { audio, bitrateSelector, favButton, pipedInstances, playButton, subtitleContainer, subtitleSelector, subtitleTrack } from "./dom";
+import { convertSStoHHMMSS, getDB, getSaved, params, parseTTML, setMetaData } from "./utils";
 import discover from "../scripts/discover";
 
 const isSafari = navigator.userAgent.indexOf('Safari') > -1 && navigator.userAgent.indexOf('Chrome') <= -1;
@@ -112,17 +112,21 @@ export default async function player(id: string | null = '') {
 
   audio.dataset.id = id;
   audio.dataset.thumbnail = data.thumbnailUrl;
-  audio.dataset.name = data.title;
+  audio.dataset.title = data.title;
   audio.dataset.author = data.uploader;
   audio.dataset.duration = convertSStoHHMMSS(data.duration);
   audio.dataset.channelUrl = data.uploaderUrl;
 
 
-  // reset fav button state
+  const db = getDB();
 
+  // reset & set favbutton state
 
-  // load related streams into discovery data after 10 seconds of playback
+  if (favButton.checked) favButton.click();
+  if (db.favorites.hasOwnProperty(id)) favButton.click();
 
-  setTimeout(discover, 1e4, data.relatedStreams, id);
+  // load related streams into discovery data after 10 seconds of constant playback
+
+  setTimeout(discover, 1e4, data.relatedStreams, id, db);
 
 }
