+++ src/scripts/main.ts
@@ -13,9 +13,8 @@ import '../components/listItem';
 import '../components/toggleSwitch';
 import { blankImage, getSaved, idFromURL, params } from '../lib/utils';
 import player from '../lib/player';
-import { audio, img } from '../lib/dom';
+import { img } from '../lib/dom';
 import { appendToQueuelist, clearQ, firstItemInQueue } from './queue';
-import { addToCollection, removeFromCollection } from './library';
 
 
 const streamQuery = params.get('s') || idFromURL(params.get('url')) || idFromURL(params.get('text'));
@@ -24,18 +23,6 @@ const streamQuery = params.get('s') || idFromURL(params.get('url')) || idFromURL
 streamQuery ? player(streamQuery) : img.src = getSaved('img') ? blankImage : '/ytify_thumbnail_min.webp';
 
 
-const favButton = <HTMLInputElement>document.getElementById('favButton');
-
-const icons = ['ri-heart-line', 'ri-heart-fill'];
-favButton.addEventListener('click', () => {
-  if (!audio.dataset.id) return;
-  favButton.checked ?
-    addToCollection('favorites', audio.dataset) :
-    removeFromCollection('favorites', <string>audio.dataset.id);
-
-  favButton.nextElementSibling?.classList.replace(icons[0], icons[1]);
-  icons.reverse();
-})
